requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
scikit-learn
newsapi-python
jinja2
